requirements.txt
appdirs==1.4.4
apturl
arabic-reshaper==2.1.1
asgiref==3.2.10
autopep8==1.5.4
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
boto3==1.26.37
botocore==1.29.37
Brlapi==0.7.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.0.8
django-autoslug-iplweb==1.9.4.dev0
django-background-tasks==1.2.5
django-chunked-upload==2.0.0
django-cities==0.5.0.6
django-cities-light==3.5.0
django-cleanup==6.0.0
django-compat==1.0.15
django-cors-headers==3.4.0
django-credit-cards==0.4.1
django-crispy-forms==1.11.2
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-environ==0.4.5
django-filter==2.4.0
django-js-asset==1.2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-multiselectfield==0.1.12
django-mysql==3.9.0
django-nine==0.2.3
django-nose==1.4.6
django-ormex==0.2.1
django-request-logging==0.7.2
django-rest-auth==0.9.5
django-rest-elasticsearch==0.4.2
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
drf-yasg==1.20.0
duplicity==0.8.12.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
entrypoints==0.3
et-xmlfile==1.0.1
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
geocoder==1.38.1
geographiclib==1.52
geolocation-python==0.2.2
geopy==2.2.0
haversine==2.7.0
html5lib==1.1
httplib2==0.14.0
idna==2.10
inflection==0.5.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.11.1
jmespath==1.0.1
jsonfield==3.1.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lml==0.0.9
lockfile==0.12.2
logger==1.4
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==1.1.1
monotonic==1.5
mysql-connector-python==8.0.29
mysqlclient==2.1.1
netifaces==0.10.4
nose==1.3.7
numpy==1.19.2
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.4
pandas==1.5.2
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
Pillow==7.2.0
progressbar2==3.47.0
protobuf==3.20.1
psycopg2-binary==2.8.5
pycairo==1.16.2
pycodestyle==2.6.0
pycountry==19.8.18
pycryptodome==3.9.9
pycryptodomex==3.9.9
pycups==1.9.73
pyexcel==0.6.6
pyexcel-io==0.6.3
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==3.12.3
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-bidi==0.4.2
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-utils==2.3.0
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
ratelim==0.1.6
razorpay==1.2.0
reportlab==3.5.56
requests==2.24.0
requests-unixsocket==0.2.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.6.0
SecretStorage==2.3.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
stripe==4.2.0
swapper==1.1.2.post1
systemd-python==234
texttable==1.6.3
toml==0.10.2
tqdm==4.43.0
typing-extensions==4.4.0
ubuntu-advantage-tools==27.11.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
usb-creator==0.3.7
virtualenv==20.4.2
wadllib==1.3.3
webencodings==0.5.1
xhtml2pdf==0.2.5
xkit==0.0.0
xlrd==1.2.0
xlwt==1.3.0
